backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
numpy==2.1.1
scipy==1.14.1
matplotlib==3.9.2
pydantic==2.9.2
python-dotenv==1.0.1
pydub==0.25.1
